Edit scss in header, refactor svg in buttons
