+++ src/shared/ui/index.ts
@@ -3,3 +3,4 @@ export { ReusableButton } from './reusable-button/index';
 export { LogoLink } from './logo-link/index';
 export { CustomIcon } from './custom-icon/index';
 export { SeparatorLine } from './separator-line/index';
+export { HeartSvg, ProfileSVG, ArrowSvg } from './svg-components';
