+++ src/shared/ui/reusable-button/styles.module.scss
@@ -7,11 +7,13 @@
     background-color: v.$purple-light;
     color: v.$dark-white;
     transition: all 0.7s ease;
+    stroke: v.$white;
 
     &:hover {
         transition: all 0.4s ease;
         background-color: v.$dark-white;
         color: v.$purple-light;
+        stroke: v.$purple-light;
     }
 
     &:active {
