+++ src/shared/ui/svg-components/ArrowSvg.tsx
@@ -1,0 +1,24 @@
+export const ArrowSvg = () => {
+    return (
+        <svg
+            width="28"
+            height="28"
+            viewBox="0 0 28 28"
+            fill="none"
+            xmlns="http://www.w3.org/2000/svg"
+        >
+            <path
+                d="M6.55469 13.5469H22.3588"
+                stroke-width="2.25773"
+                stroke-linecap="round"
+                stroke-linejoin="round"
+            />
+            <path
+                d="M14.4561 5.64453L22.3581 13.5466L14.4561 21.4487"
+                stroke-width="2.25773"
+                stroke-linecap="round"
+                stroke-linejoin="round"
+            />
+        </svg>
+    );
+};
