+++ src/shared/ui/svg-components/HeartSvg.tsx
