+++ src/shared/ui/svg-components/index.ts
@@ -1,0 +1,3 @@
+export { HeartSvg } from './HeartSvg';
+export { ProfileSVG } from './ProfileSVG';
+export { ArrowSvg } from './ArrowSvg';
